added post logic for newQuestion
