+++ src/models/Question.js
@@ -1,8 +1,10 @@
 const mongoose = require('../db/connection')
 
 const QuestionSchema = new mongoose.Schema ({
-    content: String,
-    count: Number
+    question1: String,
+    question2: String,
+    count1: Number,
+    count2: Number
 })
 
 const Question = mongoose.model("Question", QuestionSchema)
@@ -12,3 +14,4 @@ module.exports = {
 }
 
 // module.export = mongoose.model("Question", QuestionSchema)
+
